move: Move grid_convergence test to error module
